Add option to specify table env. in LaTeX output
This commit adds a command line switch (`--latex-table-environment`) for
switching between possible table environments. Currently pandoc is
hardcoded to use `longtable`, which allows for multi-page-tables. This
commit adds support for the tabularx environment (allowing columns that
span over the while remaining table-width) and gives the possibility to
easily add new environments in the future.