Removed custom Setup.hs, use build-type: simple.
The only thing we gained from the custom build was
automatic installation of the man page when using
'cabal install'.  But custom builds cause problems,
e.g., with cross-compilation.

Installation of the man page is better handled by packagers.

Note to packagers (e.g. Debian):  it may be necessary
to add a step installing the man page with the next
release.