LaTeX template: Use `footnotehyper` if available
Use the [`footnotehyper`](https://ctan.org/pkg/footnotehyper/) package if available, a rewrite of `footnote` that is compatible with `hyperref` and `babel-frenchb`. This also addresses the incompatibility with `xcolor` noted in #4861, but the new package has only been available since 2016, so this template still loads `xcolor` earlier for compatibility with older distributions. Note that the `footnote` package is part of `collection-latexrecommended` in TeX Live, so there shouldn't be any situations where it isn't available.