Writers.Shared.metaValueToJSON: use Number Values for integers.
Pandoc's MetaValue doesn't have a distinguished number type,
so numbers are put in MetaStrings.  If the MetaString consists
entirely of digits, we convert it to a Number.

We should probably consider adding a MetaNumber constructor
to MetaValue, for better round-tripping with JSON etc.

This change aids round-tripping in ipynb metadata fields,
like `toc_depth`.