LaTeX template: Ensure correct heading/table order
Improve the workaround for #1658, adapting a solution by @u-fischer in <https://github.com/latex3/latex2e/issues/131> that works whether or not the `indent` variable is enabled. Rename the previous approach under the `subparagraph` variable as `block-headings`. This is no longer enabled by default, as it gives headings 3–5 an identical appearance. Provide an example in the manual of reformatting the appearance of headings more thoroughly using KOMA-Script. Closes #5365.