Change optMetadataFile type from Maybe to List
Changed optMetadataFile from Maybe FilePath to [FilePath]. This allows
for multiple YAML metadata files to be added. The new default value has
been changed from Nothing to [].

To account for this change in Text.Pandoc.App, metaDataFromFile now
operates on two mapM calls (for readFileLazy and yamlToMeta) and a fold.

Added a test (command/5700.md) which tests this functionality and
updated MANUAL.txt, as per the contributing guidelines.