Add title field to 5700-metadta-file-2.yml
Suggested by @mb21 for testing the behavior of how the files are merged.

With the current behavior, using foldr1 (<>), values within files
specified first will be used over those in later files. (If the reverse
of this behavior would be preferred, it should be fixed by changing
foldr1 to foldl1.)