Use the text branch of pandoc-types and texmath, switch to Legacy.*
As part of the switch to Text, imports of Pandoc.Builder and
Pandoc.Definition were changed to their Legacy
counterparts. Compatibility functions were added to modules using
texmath. These will all be removed as refactoring progresses.

The stack.yaml extra-deps for pandoc-types and texmath now use the
latest commit in the text branch of those packages.