Fix regression in `makeSections`.
Previously hierarchicalize (the ancestor of `makeSections`)
would put header attributes on the containing Div.  In 2.8
this behavior changed, which broke some tools depending
on pandoc.  Here we roll back this change, so that attributes
again migrate from the header to the containing Div when
`makeSections` is run.  Note that attributes are retained
on the header as well (unlike before) -- with the exception
of the `id` attribute, which of course cannot be duplicated.

Note that this is an empty commit.  The previous
commit, 8d0033111, was mistakenly recorded as an
amendment to the --toc-depth fix, so the commit message
for that is wrong; this message correctly describes
the change in 8d0033111.

Closes #5965.