Ensure that later default file values for `variable` replace earlier ones.
The semigroup instance for doctemplates Context does a left-biased
union on the underlying Map.  That means that if you union
`{a: 1}` and `{a: 2}`, you get `{a: 1}`.  This commit causes
pandoc to do the operation in the opposite order, so that later
default files take precedence in the values they assign to
keys.

See #5988.