Docx writer: fix regression with Compact style on tight lists.
Starting in 2.8, the docx writer no longer distinguishes
between tight and loose lists, since the Compact style is
omitted.

This is a side-effect of the fix to #5670, as explained
in the changelog:

    + Preserve built-in styles in DOCX with custom style (Ben Steinberg,
      #5670).  This change prevents custom styles on divs and spans
      from overriding styles on certain elements inside them, like
      headings, blockquotes, and links. On those elements, the
      "native" style is required for the element to display correctly.
      This change also allows nesting of custom styles; in order to do so,
      it removes the default "Compact" style applied to Plain blocks,
      except when inside a table.

This patch fixes the problem by extending the exception currently
offered to Plain blocks inside tables to Plain blocks inside list
items.

Closes #6072.