tools: delete Lua documentation helper cruft
The tools were useful to generate documentation from the Lua sources.
However, most Lua functions are defined in Haskell, and the even
documentation of functions written in Lua has been updated manually for
quite a while now. The unused helper scripts, which have bit-rot to the
point that they now produce broken links, are therefore removed.