Update revealjs template with newly available options
I wanted to use an option and realized that the default pandoc template
didn't pass it through, so I went through and found all the options that
Reveal.js advertises but that Pandoc's template doesn't support.

<https://github.com/hakimel/reveal.js#configuration>

I also noticed that rollingLinks isn't a setting anymore, but it's fine
to keep it in the template because people might be using an old version
of Reveal.js (via `revealjs-url`) that still has this option.