Use addchap instead of chapter* for bookscr
Using chapter* in bookscr makes the running head incorrect in the chapter (the previous running head is used). KomaScript provides a special function for this usecase: addchap. It generates a new chapter, adds proper running head and also generates a toc item, so "addcontentsline" case be skipped in this case.