Fix .gitattributes for Win OS
Enforce the EditorConfig EOL normalization rules for text files to LF,
inside `.gitattributes`, for contributors working with Windows OS:

    *.* text eol=lf

Since the current `.editorconfig` enforces LF end-of-lines on all files,
Windows users working with an EditorConfig-compatible editor are unable
to commit changes when Git is configured with the (recommended) settings
`core.eol = native` and `core.autocrlf = true` , because all text files
are normalized by the editor to LF EOL at save time, but Git expects
native EOL (CRLF) at commit time.

This is a quick fix to allow editing text documents, but the repository
settings for EditorConfig and Git attributes should be revisited to
ensure an optimal cross-platform collaboration.