added support for writing nested tables to asciidoc
Added field to WriterState that denotes the current nesting level for traversing tables.
Depending on the value of that field nested tables are recognized and written.
Asciidoc supports one level of nesting. If deeper tables are to be written, they are
written as one level nested tables and enclosed in a comment explaining that limitation.