Lua: disable implicit execution of user-provided init file
Pandoc used to load the file `init.lua` from a user data directory
during Lua initialization. This feature has been removed, as it was in
conflict with pandoc's philosophy of only loading files that have been
passed explicitly. This also solves a slight security issue, as write
permissions to a user's data directory automatically resulted in the
power to run arbitrary Lua commands.